Перенесены классы из MovieQuiz в отдельные группы. Создана фабрика вопросов
