+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+//Структура вопроса
+struct QuizQuestion{
+    let image:String
+    let text:String
+    let correctAnswer:Bool
+}
