+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -26,6 +26,9 @@ final class MovieQuizViewController: UIViewController {
 
     private var currentQuestionIndex:Int = 0
     private var correctAnswers:Int = 0
+    private let questionsAmount:Int = 10
+    private var questionFactory:QuestionFactory = QuestionFactory()
+    private var qurentQuestion: QuizQuestion?
     //private let currentQuestion = questions[currentQuestionIndex]
     
     override func viewDidLoad() {
@@ -34,70 +37,6 @@ final class MovieQuizViewController: UIViewController {
         let currentQuestion = questions[currentQuestionIndex]
         let questionStep = convert(model: currentQuestion)
         show(quiz: questionStep)
-    }
-    
-    private let questions: [QuizQuestion] = [
-            QuizQuestion(
-                image: "The Godfather",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Dark Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Kill Bill",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Avengers",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Deadpool",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Green Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Old",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "The Ice Age Adventures of Buck Wild",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Tesla",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Vivarium",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false)
-        ]
-    
-    //Структура вопроса
-    struct QuizQuestion{
-        let image:String
-        let text:String
-        let correctAnswer:Bool
-    }
-    
-    //Вопрос показан
-    struct QuizStepViewModel{
-        let image:UIImage
-        let question:String
-        let questionNumber:String
-    }
-    
-    //Результат квиза
-    struct QuizResultsViewModel{
-        let title:String
-        let text:String
-        let buttonText:String
     }
     
     //Конвертация View model
